requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
slack_sdk
aiohttp
matplotlib
together
langgraph
